Create README.md
The Tip Calculator is a Project aimed at providing a helpful resource for individuals seeking assistance with calculating tips. The repository contains a concise and efficient Python program that calculates the tip amount and the total bill based on user input. It serves as a convenient tool for those looking to quickly determine the appropriate tip based on the total amount and desired tip percentage. The repository encourages collaboration and provides an avenue for interested individuals to access and contribute to the project, fostering a community of users supporting each other in the realm of tipping etiquette.